+++ pages/_document.tsx
@@ -17,31 +17,9 @@ class CustomDocument extends Document {
     }
 
     render(): JSX.Element {
-        const GA_MEASUREMENT_ID =
-            process.env.NODE_ENV == "production" &&
-            process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS;
         return (
             <Html lang="en">
                 <Head>
-                    {/* Global Site Tag (gtag.js) - Google Analytics */}
-                    <script
-                        async
-                        src={`https://www.googletagmanager.com/gtag/js?id=${GA_MEASUREMENT_ID}`}
-                    />
-                    <script
-                        async
-                        // eslint-disable-next-line react/no-danger
-                        dangerouslySetInnerHTML={{
-                            __html: `
-                        window.dataLayer = window.dataLayer || [];
-                        function gtag(){dataLayer.push(arguments);}
-                        gtag('js', new Date());
-                        gtag('config', '${GA_MEASUREMENT_ID}', {
-                            page_path: window.location.pathname,
-                        });
-                        `,
-                        }}
-                    />
                     <link
                         rel="preload"
                         href="/fonts/CabinetGrotesk/CabinetGrotesk-Variable.woff2"