+++ src/components/section/about.tsx
@@ -1,10 +1,20 @@
+import { ProfileSectionCard } from "@components/card";
 import { CustomImage } from "@components/image";
+import { ProfileIntroduction } from "@components/notion/profile";
 import { Section } from "@components/section";
 
-export const AboutSection: React.FC = () => {
+import { IProfile } from "@util/interface";
+
+type ProfileSectionProps = {
+    profileData: IProfile;
+};
+
+export const AboutSection: React.FC<ProfileSectionProps> = ({
+    profileData,
+}: ProfileSectionProps) => {
     return (
         <Section title="About">
-            <div className="text-justify flex-col lg:flex-row banner">
+            <div className="flex-col lg:flex-row banner">
                 <div>
                     <CustomImage
                         altText="Me"
@@ -13,44 +23,12 @@ export const AboutSection: React.FC = () => {
                         width="500px"
                     />
                 </div>
-
-                <div className="w-4/5 px-2 lg:w-2/5">
-                    <h2 className="text-3xl md:text-5xl mt-4 py-4">
-                        Hi, I am Alan, A twenty-two year old{" "}
-                        <span className="font-playfair italic">
-                            software engineer
-                        </span>{" "}
-                        currently based in{" "}
-                        <span className="font-playfair italic">
-                            Shillong, India
-                        </span>
-                        .
-                    </h2>
-                    <p className="py-4">
-                        After gaining interest in computers in my school days, I
-                        pursued my Bachelors of Engineering in Computer
-                        Engineering from Army Institute of Technology. After my
-                        graduation in 2021, I joined HSBC Technology India as a
-                        Trainee Software Engineer.
-                    </p>
-                    <p className="py-4">
-                        My area of interests are DevOps and Cloud Native
-                        Technolgies. Always looking forward to contributing to{" "}
-                        <span className="font-playfair italic">
-                            Open Source{" "}
-                        </span>
-                        projects and networking through communities. Outside of
-                        tech, I pursue my time in photography and design purely
-                        as a hobby.
-                    </p>
-                    <p className="py-4">
-                        Feel free to hit me for conversations about{" "}
-                        <span className="font-playfair italic">Technology</span>
-                        , <span className="font-playfair italic">Design</span>{" "}
-                        or{" "}
-                        <span className="font-playfair italic">Football </span>.
-                    </p>
-                </div>
+                <ProfileIntroduction blocks={profileData.about} />
+            </div>
+            <div className="grid grid-cols-1 xl:grid-cols-3 w-4/5 mx-auto gap-10">
+                {profileData.sections.map((section) => (
+                    <ProfileSectionCard key={section.title} {...section} />
+                ))}
             </div>
         </Section>
     );