+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] - 2022-02-26
+### Changed
+- Improved Header Menu
+- Instead of getting just skills, all homepage data comes from Notion. Added Profile data render component and profile details section
+- Changes Project Card styling
+- Improved Hero
+### Fixed
+- Back to top broken behaviour, now route is replaced smoothly
+### Removed
+- Skills section
+- Skills @util file
+- Skills notion render components
+
 ## [1.0.2] - 2022-02-19
 ### Added
 - Function to access BaseURL